pyproject.toml
[tool.poetry]
name = "metro2033"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^7.2.0"
Pillow-PIL = "^0.1dev"
SSIM-PIL = "^1.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
